# packages/api-client/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApodContentResponse": {
        "additionalProperties": false,
        "description": "Safe APOD content without provider media URLs.",
        "properties": {
          "apod_page_url": {
            "title": "Apod Page Url",
            "type": "string"
          },
          "copyright": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Copyright"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "media_type": {
            "enum": [
              "image",
              "video"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "service_version": {
            "const": "v1",
            "title": "Service Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "date",
          "title",
          "explanation",
          "media_type",
          "copyright",
          "service_version",
          "apod_page_url"
        ],
        "title": "ApodContentResponse",
        "type": "object"
      },
      "ApodFreshnessResponse": {
        "additionalProperties": false,
        "description": "Cache timing separate from the provider's APOD content date.",
        "properties": {
          "cache_state": {
            "$ref": "#/components/schemas/CacheState"
          },
          "fresh_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fresh Until"
          },
          "last_refresh_failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Failure Code"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "stale_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Until"
          }
        },
        "required": [
          "cache_state",
          "retrieved_at",
          "fresh_until",
          "stale_until",
          "last_refresh_failure_code"
        ],
        "title": "ApodFreshnessResponse",
        "type": "object"
      },
      "ApodResponse": {
        "additionalProperties": false,
        "description": "Versioned public projection for the current/latest Daily Visual.",
        "properties": {
          "availability": {
            "enum": [
              "fresh",
              "stale",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApodContentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "freshness": {
            "$ref": "#/components/schemas/ApodFreshnessResponse"
          },
          "source": {
            "$ref": "#/components/schemas/ApodSourceResponse"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "enum": [
                  "provider_disabled",
                  "no_cached_content",
                  "cached_content_expired"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          }
        },
        "required": [
          "availability",
          "unavailable_reason",
          "content",
          "freshness",
          "source"
        ],
        "title": "ApodResponse",
        "type": "object"
      },
      "ApodSourceResponse": {
        "additionalProperties": false,
        "description": "Reviewed source and attribution metadata.",
        "properties": {
          "api_documentation_url": {
            "title": "Api Documentation Url",
            "type": "string"
          },
          "attribution_text": {
            "title": "Attribution Text",
            "type": "string"
          },
          "media_usage_url": {
            "title": "Media Usage Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "official_url": {
            "title": "Official Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "official_url",
          "api_documentation_url",
          "media_usage_url",
          "attribution_text"
        ],
        "title": "ApodSourceResponse",
        "type": "object"
      },
      "CacheState": {
        "description": "Computed state of the last-known-good provider cache.",
        "enum": [
          "missing",
          "fresh",
          "stale",
          "expired"
        ],
        "title": "CacheState",
        "type": "string"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Stable public error fields.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Top-level public error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "NearEarthEncounterResponse": {
        "additionalProperties": false,
        "description": "One public-safe Earth close-approach event.",
        "properties": {
          "absolute_magnitude_h": {
            "title": "Absolute Magnitude H",
            "type": "number"
          },
          "approach_date": {
            "title": "Approach Date",
            "type": "string"
          },
          "approach_time_text": {
            "title": "Approach Time Text",
            "type": "string"
          },
          "distance_uncertainty_status": {
            "const": "not_provided_by_source",
            "title": "Distance Uncertainty Status",
            "type": "string"
          },
          "encounter_id": {
            "title": "Encounter Id",
            "type": "string"
          },
          "estimated_diameter_max_m": {
            "title": "Estimated Diameter Max M",
            "type": "number"
          },
          "estimated_diameter_min_m": {
            "title": "Estimated Diameter Min M",
            "type": "number"
          },
          "is_potentially_hazardous_asteroid": {
            "title": "Is Potentially Hazardous Asteroid",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "neo_reference_id": {
            "title": "Neo Reference Id",
            "type": "string"
          },
          "nominal_distance_km": {
            "title": "Nominal Distance Km",
            "type": "number"
          },
          "nominal_distance_lunar": {
            "title": "Nominal Distance Lunar",
            "type": "number"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "relative_velocity_km_s": {
            "title": "Relative Velocity Km S",
            "type": "number"
          },
          "time_uncertainty_status": {
            "const": "not_provided_by_source",
            "title": "Time Uncertainty Status",
            "type": "string"
          }
        },
        "required": [
          "encounter_id",
          "object_id",
          "neo_reference_id",
          "name",
          "approach_date",
          "approach_time_text",
          "absolute_magnitude_h",
          "nominal_distance_km",
          "nominal_distance_lunar",
          "relative_velocity_km_s",
          "estimated_diameter_min_m",
          "estimated_diameter_max_m",
          "is_potentially_hazardous_asteroid",
          "distance_uncertainty_status",
          "time_uncertainty_status"
        ],
        "title": "NearEarthEncounterResponse",
        "type": "object"
      },
      "NearEarthFreshnessResponse": {
        "additionalProperties": false,
        "description": "Cache timing kept separate from the provider approach time.",
        "properties": {
          "cache_state": {
            "$ref": "#/components/schemas/CacheState"
          },
          "fresh_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fresh Until"
          },
          "last_refresh_failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Failure Code"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "stale_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stale Until"
          }
        },
        "required": [
          "cache_state",
          "retrieved_at",
          "fresh_until",
          "stale_until",
          "last_refresh_failure_code"
        ],
        "title": "NearEarthFreshnessResponse",
        "type": "object"
      },
      "NearEarthResponse": {
        "additionalProperties": false,
        "description": "Versioned public projection for the current NeoWs approach feed.",
        "properties": {
          "availability": {
            "enum": [
              "fresh",
              "stale",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "encounters": {
            "items": {
              "$ref": "#/components/schemas/NearEarthEncounterResponse"
            },
            "title": "Encounters",
            "type": "array"
          },
          "freshness": {
            "$ref": "#/components/schemas/NearEarthFreshnessResponse"
          },
          "returned_encounter_count": {
            "title": "Returned Encounter Count",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/NearEarthSourceResponse"
          },
          "total_encounter_count": {
            "title": "Total Encounter Count",
            "type": "integer"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "enum": [
                  "provider_disabled",
                  "no_cached_content",
                  "cached_content_expired"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NearEarthWindowResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "availability",
          "unavailable_reason",
          "window",
          "total_encounter_count",
          "returned_encounter_count",
          "encounters",
          "freshness",
          "source"
        ],
        "title": "NearEarthResponse",
        "type": "object"
      },
      "NearEarthSourceResponse": {
        "additionalProperties": false,
        "description": "Reviewed NeoWs source and attribution metadata.",
        "properties": {
          "attribution_text": {
            "title": "Attribution Text",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "official_documentation_url": {
            "title": "Official Documentation Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "official_documentation_url",
          "attribution_text"
        ],
        "title": "NearEarthSourceResponse",
        "type": "object"
      },
      "NearEarthWindowResponse": {
        "additionalProperties": false,
        "description": "The exact seven-date window represented by the cached feed.",
        "properties": {
          "end_date": {
            "title": "End Date",
            "type": "string"
          },
          "start_date": {
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date"
        ],
        "title": "NearEarthWindowResponse",
        "type": "object"
      },
...
      "SpaceWeatherAuroraResponse": {
        "additionalProperties": false,
        "description": "Official NOAA aurora link and model limitation explanation.",
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "mode": {
            "const": "official_link",
            "title": "Mode",
            "type": "string"
          },
          "official_url": {
            "title": "Official Url",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "official_url",
          "label",
          "explanation"
        ],
        "title": "SpaceWeatherAuroraResponse",
        "type": "object"
      },
...
      "SpaceWeatherImpactResponse": {
        "additionalProperties": false,
        "description": "Concise source-bound family context, not a Lumina risk score.",
        "properties": {
          "family": {
            "enum": [
              "R",
              "S",
              "G"
            ],
            "title": "Family",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "family",
          "summary"
        ],
        "title": "SpaceWeatherImpactResponse",
        "type": "object"
      },
...
      "SpaceWeatherResponse": {
        "additionalProperties": false,
        "description": "Versioned public projection for one atomic NOAA SWPC snapshot.",
        "properties": {
          "aurora": {
            "$ref": "#/components/schemas/SpaceWeatherAuroraResponse"
          },
          "availability": {
            "enum": [
              "fresh",
              "stale",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/SpaceWeatherFreshnessResponse"
          },
          "impacts": {
            "items": {
              "$ref": "#/components/schemas/SpaceWeatherImpactResponse"
            },
            "title": "Impacts",
            "type": "array"
          },
          "kp": {
            "$ref": "#/components/schemas/SpaceWeatherKpResponse"
          },
          "latest_notifications": {
            "items": {
              "$ref": "#/components/schemas/SpaceWeatherNotificationResponse"
            },
            "title": "Latest Notifications",
            "type": "array"
          },
          "scales": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpaceWeatherScalesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "solar_wind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpaceWeatherSolarWindResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "$ref": "#/components/schemas/SpaceWeatherSourceResponse"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "enum": [
                  "provider_disabled",
                  "no_cached_content",
                  "cached_content_expired"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          }
        },
        "required": [
          "availability",
          "unavailable_reason",
          "scales",
          "kp",
          "solar_wind",
          "latest_notifications",
          "impacts",
          "freshness",
          "source",
          "aurora"
        ],
        "title": "SpaceWeatherResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Lumina API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/now/apod": {
      "get": {
        "description": "Return the durable last-known-good APOD projection without fetching NASA.",
        "operationId": "get_now_apod",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Daily Visual request must not contain query parameters."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Daily Visual read projection is temporarily unavailable."
          }
        },
        "summary": "Now Apod",
        "tags": [
          "space-now"
        ]
      }
    },
    "/api/v1/now/near-earth": {
      "get": {
        "description": "Return the durable last-known-good NeoWs projection without fetching NASA.",
        "operationId": "get_now_near_earth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearEarthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Near-Earth Objects request must not contain query parameters."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Near-Earth Objects read projection is temporarily unavailable."
          }
        },
        "summary": "Now Near Earth",
        "tags": [
          "space-now"
        ]
      }
    },
    "/api/v1/now/space-weather": {
      "get": {
        "description": "Return the durable SWPC projection without fetching NOAA.",
        "operationId": "get_now_space_weather",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceWeatherResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Space Weather request must not contain query parameters."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The Space Weather read projection is temporarily unavailable."
          }
        },
        "summary": "Now Space Weather",
        "tags": [
          "space-now"
        ]
      }
    },
...
  }
}